ml_service/requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.11
